extract getActiveStudentCourses into object-student
